// packages/shared/chat-ui/src/types.ts
export interface Attachment {
  name: string
  type: string
  dataUrl?: string
  content?: string
}